1. Any test that references nctestserver/NC_findtestserver    should be under ENABLE_DAP_REMOTE_TESTS.    Fixed to make sure that this is so.    Also attempted to fix ncdap_test/CMakeLists.txt,    but probably got it wrong.    HT to Nico Schlomer. 2. Attempted to reduce the number of conversion errors    when -Wconversion is set. Fixed oc2, but    rest of netcdf remains to be done.    HT to Nico Schlomer. 3. When doing #2, I discovered an error in ncgen.y    that has remained hidden. This required some other    test case fixes.

+++ ncdump/bom.c
@@ -6,6 +6,7 @@
 #include <config.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <string.h>
 
 /* BOM Sequences */
 static char* U8   = "\xEF\xBB\xBF";    /* UTF-8 */
@@ -18,7 +19,7 @@ int
 main(int argc, char** argv)
 {
     char* bom = U8;
-    int bomlen = 3;
+    size_t bomlen = 3;
     if(argc > 1 && strlen(argv[1]) > 0) {
 	char* which = argv[1];
 	switch (which[0]) {
@@ -30,4 +31,3 @@ main(int argc, char** argv)
     fwrite(bom,1,bomlen,stdout);
     exit(0);
 }
-
